End of 10 Jun 2019 II

+++ src/test/java/com/test/oic/OICvalidatePolicyType.java
@@ -9,7 +9,7 @@ import org.testng.annotations.Test;
 public class OICvalidatePolicyType extends BaseOIC {
 
     @Test(dataProvider = "policyTypeValidData")
-    public void validateValidPolicyType(double sellNet, String policyType, String carCode, int engineSize, int driverAge, int carAge, int sumInsured, int carGroup, int injuryTPPerson, int injuryTPTime, int injuryTPDamage, int seat, int coverageLifeloss, int coverageMedical, int coverageBailbond, int deductible) {
+    public void validateValidPolicyType(float sellNet, String policyType, String carCode, int engineSize, int driverAge, int carAge, int sumInsured, int carGroup, int injuryTPPerson, int injuryTPTime, int injuryTPDamage, int seat, int coverageLifeloss, int coverageMedical, int coverageBailbond, int deductible) {
         //Arrange
         var request = RestAssured.given().spec(requestSpec);
         assignParameter(sellNet, policyType, carCode, engineSize, driverAge, carAge, sumInsured, carGroup, injuryTPPerson, injuryTPTime, injuryTPDamage, seat, coverageLifeloss, coverageMedical, coverageBailbond, deductible);
